+++ packages/frontend/src/rtl-tests/community.join.test.tsx
@@ -312,6 +312,7 @@ describe('User', () => {
         "Modals/openModal",
         "Identity/registerUsername",
         "Identity/registerCertificate",
+        "Errors/handleError",
         "Errors/addError",
       ]
     `)