+++ packages/nectar/src/sagas/errors/handleErrors/handleErrors.saga.ts
@@ -23,14 +23,28 @@ function* retryRegistration(communityId: string) {
   log(`registering certificate for community ${communityId} failed, trying again`)
 }
 
+let registrationAttempts: number = 0
+
 export function* handleErrorsSaga(
   action: PayloadAction<ReturnType<typeof errorsActions.addError>['payload']>
 ): Generator {
   const error: ErrorPayload = action.payload
   if (error.type === SocketActionTypes.REGISTRAR) {
-    if (error.code === ErrorCodes.NOT_FOUND) {
-      yield* call(delay, 15000)
-      yield* call(retryRegistration, error.community)
+    if (error.code === ErrorCodes.FORBIDDEN) {
+      yield* put(errorsActions.addError(error))
     }
+    if (error.code === ErrorCodes.NOT_FOUND || error.code === ErrorCodes.SERVER_ERROR || error.code === ErrorCodes.SERVICE_UNAVAILABLE) {
+      // 7 retries guarantees 99.9% chance of registering username if registrar is online.
+      if (registrationAttempts < 7) {
+        yield* call(delay, 5000)
+        registrationAttempts++
+        yield* call(retryRegistration, error.community)
+      } else {
+        yield* put(errorsActions.addError(error))
+        registrationAttempts = 0
+      }
+    }
+  } else {
+    yield* put(errorsActions.addError(error))
   }
 }