+++ packages/e2e-tests/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.4-alpha.0](https://github.com/ZbayApp/monorepo/compare/e2e-tests@0.4.3-alpha.0...e2e-tests@0.4.4-alpha.0) (2022-04-15)
+
+**Note:** Version bump only for package e2e-tests
+
+
+
+
+
+## [0.4.3-alpha.0](https://github.com/ZbayApp/monorepo/compare/e2e-tests@0.4.2-alpha.0...e2e-tests@0.4.3-alpha.0) (2022-04-15)
+
+**Note:** Version bump only for package e2e-tests
+
+
+
+
+
 ## [0.4.2-alpha.0](https://github.com/ZbayApp/monorepo/compare/e2e-tests@0.4.0-alpha.3...e2e-tests@0.4.2-alpha.0) (2022-04-15)
 
 **Note:** Version bump only for package e2e-tests